funzia, con variabili corrette

+++ src/mp/progetto/DiscountStrategy.java
@@ -1,5 +1,5 @@
 package mp.progetto;
 
 public interface DiscountStrategy {
-	int applyDiscount(int originalPrice);
+	double applyDiscount(double originalPrice);
 }
